refactor(tasks): remove dead scheduler code
