+++ mock/task_service.go
@@ -130,8 +130,6 @@ func (s *TaskService) ForceRun(ctx context.Context, taskID influxdb.ID, schedule
 }
 
 type TaskControlService struct {
-	CreateNextRunFn    func(ctx context.Context, taskID influxdb.ID, now int64) (backend.RunCreation, error)
-	NextDueRunFn       func(ctx context.Context, taskID influxdb.ID) (int64, error)
 	CreateRunFn        func(ctx context.Context, taskID influxdb.ID, scheduledFor time.Time, runAt time.Time) (*influxdb.Run, error)
 	CurrentlyRunningFn func(ctx context.Context, taskID influxdb.ID) ([]*influxdb.Run, error)
 	ManualRunsFn       func(ctx context.Context, taskID influxdb.ID) ([]*influxdb.Run, error)
@@ -141,12 +139,6 @@ type TaskControlService struct {
 	AddRunLogFn        func(ctx context.Context, taskID, runID influxdb.ID, when time.Time, log string) error
 }
 
-func (tcs *TaskControlService) CreateNextRun(ctx context.Context, taskID influxdb.ID, now int64) (backend.RunCreation, error) {
-	return tcs.CreateNextRunFn(ctx, taskID, now)
-}
-func (tcs *TaskControlService) NextDueRun(ctx context.Context, taskID influxdb.ID) (int64, error) {
-	return tcs.NextDueRunFn(ctx, taskID)
-}
 func (tcs *TaskControlService) CreateRun(ctx context.Context, taskID influxdb.ID, scheduledFor time.Time, runAt time.Time) (*influxdb.Run, error) {
 	return tcs.CreateRunFn(ctx, taskID, scheduledFor, runAt)
 }
