+++ task/mock/task_control_service.go
@@ -2,7 +2,6 @@ package mock
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"sort"
 	"sync"
@@ -11,7 +10,6 @@ import (
 	"github.com/influxdata/influxdb"
 	"github.com/influxdata/influxdb/snowflake"
 	"github.com/influxdata/influxdb/task/backend"
-	cron "gopkg.in/robfig/cron.v2"
 )
 
 var idgen = snowflake.NewDefaultIDGenerator()
@@ -23,7 +21,7 @@ type TaskControlService struct {
 	runs map[influxdb.ID]map[influxdb.ID]*influxdb.Run
 
 	// Map of stringified, concatenated task and platform ID, to runs that have been created.
-	created map[string]backend.QueuedRun
+	created map[string]*influxdb.Run
 
 	// Map of stringified task ID to task meta.
 	tasks      map[influxdb.ID]*influxdb.Task
@@ -40,7 +38,7 @@ func NewTaskControlService() *TaskControlService {
 		runs:             make(map[influxdb.ID]map[influxdb.ID]*influxdb.Run),
 		finishedRuns:     make(map[influxdb.ID]*influxdb.Run),
 		tasks:            make(map[influxdb.ID]*influxdb.Task),
-		created:          make(map[string]backend.QueuedRun),
+		created:          make(map[string]*influxdb.Run),
 		totalRunsCreated: make(map[influxdb.ID]int),
 	}
 }
@@ -56,103 +54,6 @@ func (d *TaskControlService) SetTask(task *influxdb.Task) {
 
 func (d *TaskControlService) SetManualRuns(runs []*influxdb.Run) {
 	d.manualRuns = runs
-}
-
-// CreateNextRun creates the next run for the given task.
-// Refer to the documentation for SetTaskPeriod to understand how the times are determined.
-func (d *TaskControlService) CreateNextRun(ctx context.Context, taskID influxdb.ID, now int64) (backend.RunCreation, error) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-	if !taskID.Valid() {
-		return backend.RunCreation{}, errors.New("invalid task id")
-	}
-	tid := taskID
-
-	task, ok := d.tasks[tid]
-	if !ok {
-		panic(fmt.Sprintf("meta not set for task with ID %s", tid))
-	}
-
-	if len(d.manualRuns) != 0 {
-		run := d.manualRuns[0]
-		d.manualRuns = d.manualRuns[1:]
-		runs, ok := d.runs[tid]
-		if !ok {
-			runs = make(map[influxdb.ID]*influxdb.Run)
-		}
-		runs[run.ID] = run
-		d.runs[task.ID] = runs
-		next, _ := d.nextDueRun(ctx, taskID)
-		rc := backend.RunCreation{
-			Created: backend.QueuedRun{
-				TaskID: task.ID,
-				RunID:  run.ID,
-				Now:    run.ScheduledFor.Unix(),
-			},
-			NextDue:  next,
-			HasQueue: len(d.manualRuns) != 0,
-		}
-		d.created[tid.String()+rc.Created.RunID.String()] = rc.Created
-		d.totalRunsCreated[taskID]++
-		return rc, nil
-	}
-
-	rc, err := d.createNextRun(task, now)
-	if err != nil {
-		return backend.RunCreation{}, err
-	}
-	rc.Created.TaskID = taskID
-	d.created[tid.String()+rc.Created.RunID.String()] = rc.Created
-	d.totalRunsCreated[taskID]++
-	return rc, nil
-}
-
-func (t *TaskControlService) createNextRun(task *influxdb.Task, now int64) (backend.RunCreation, error) {
-	sch, err := cron.Parse(task.EffectiveCron())
-	if err != nil {
-		return backend.RunCreation{}, err
-	}
-
-	latest := task.LatestCompleted.Unix()
-
-	for _, r := range t.runs[task.ID] {
-		if r.ScheduledFor.Unix() > latest {
-			latest = r.ScheduledFor.Unix()
-		}
-
-	}
-
-	nextScheduled := sch.Next(time.Unix(latest, 0))
-	nextScheduledUnix := nextScheduled.Unix()
-
-	offset := int64(0)
-	if task.Offset != 0 {
-		offset = task.Offset.Nanoseconds()
-	}
-
-	if dueAt := nextScheduledUnix + int64(offset); dueAt > now {
-		return backend.RunCreation{}, influxdb.ErrRunNotDueYet(dueAt)
-	}
-
-	runID := idgen.ID()
-	runs, ok := t.runs[task.ID]
-	if !ok {
-		runs = make(map[influxdb.ID]*influxdb.Run)
-	}
-	runs[runID] = &influxdb.Run{
-		ID:           runID,
-		ScheduledFor: nextScheduled,
-	}
-	t.runs[task.ID] = runs
-
-	return backend.RunCreation{
-		Created: backend.QueuedRun{
-			RunID: runID,
-			Now:   nextScheduledUnix,
-		},
-		NextDue:  sch.Next(nextScheduled).Unix() + offset,
-		HasQueue: false,
-	}, nil
 }
 
 func (t *TaskControlService) CreateRun(_ context.Context, taskID influxdb.ID, scheduledFor time.Time, runAt time.Time) (*influxdb.Run, error) {
@@ -228,41 +129,6 @@ func (t *TaskControlService) ManualRuns(ctx context.Context, taskID influxdb.ID)
 	return []*influxdb.Run{}, nil
 }
 
-// NextDueRun returns the Unix timestamp of when the next call to CreateNextRun will be ready.
-// The returned timestamp reflects the task's offset, so it does not necessarily exactly match the schedule time.
-func (d *TaskControlService) NextDueRun(ctx context.Context, taskID influxdb.ID) (int64, error) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	return d.nextDueRun(ctx, taskID)
-}
-
-func (d *TaskControlService) nextDueRun(ctx context.Context, taskID influxdb.ID) (int64, error) {
-	task := d.tasks[taskID]
-	sch, err := cron.Parse(task.EffectiveCron())
-	if err != nil {
-		return 0, err
-	}
-
-	latest := task.LatestCompleted.Unix()
-
-	for _, r := range d.runs[task.ID] {
-		if r.ScheduledFor.Unix() > latest {
-			latest = r.ScheduledFor.Unix()
-		}
-	}
-
-	nextScheduled := sch.Next(time.Unix(latest, 0))
-	nextScheduledUnix := nextScheduled.Unix()
-
-	offset := int64(0)
-	if task.Offset != 0 {
-		offset = task.Offset.Nanoseconds()
-	}
-
-	return nextScheduledUnix + int64(offset), nil
-}
-
 // UpdateRunState sets the run state at the respective time.
 func (d *TaskControlService) UpdateRunState(ctx context.Context, taskID, runID influxdb.ID, when time.Time, state backend.RunStatus) error {
 	d.mu.Lock()
@@ -299,11 +165,11 @@ func (d *TaskControlService) AddRunLog(ctx context.Context, taskID, runID influx
 	return nil
 }
 
-func (d *TaskControlService) CreatedFor(taskID influxdb.ID) []backend.QueuedRun {
+func (d *TaskControlService) CreatedFor(taskID influxdb.ID) []*influxdb.Run {
 	d.mu.Lock()
 	defer d.mu.Unlock()
 
-	var qrs []backend.QueuedRun
+	var qrs []*influxdb.Run
 	for _, qr := range d.created {
 		if qr.TaskID == taskID {
 			qrs = append(qrs, qr)
@@ -325,10 +191,10 @@ func (d *TaskControlService) TotalRunsCreatedForTask(taskID influxdb.ID) int {
 // If the expected number isn't found in time, it returns an error.
 //
 // Because the scheduler and executor do a lot of state changes asynchronously, this is useful in test.
-func (d *TaskControlService) PollForNumberCreated(taskID influxdb.ID, count int) ([]backend.QueuedRun, error) {
+func (d *TaskControlService) PollForNumberCreated(taskID influxdb.ID, count int) ([]*influxdb.Run, error) {
 	const numAttempts = 50
 	actualCount := 0
-	var created []backend.QueuedRun
+	var created []*influxdb.Run
 	for i := 0; i < numAttempts; i++ {
 		time.Sleep(2 * time.Millisecond) // we sleep even on first so it becomes more likely that we catch when too many are produced.
 		created = d.CreatedFor(taskID)
